requirements.txt
Flask>=2.0
Flask-WTF>=1.0
python-dotenv>=0.21
psutil>=5.9
docker>=6.1